loops and file handling
I found a list of universities and their nicknames on wikipedia.  I wanted to revise it so that there was one nickname and one university per row. So I copied it into a file, changed the commas to pipes (removing those for things like University California, Berkeley), and wrote this for it. This took substantially longer to write than just to clean it up by hand, but it was an opportunity to practice.